Ajout PARTIEL de la création de l'utilisateur et de l'authentification

+++ EDT_back_api/api/src/app/base/models/entities/base_entity.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from sqlalchemy.orm import declarative_base
-from sqlalchemy import Column, BigInteger, String, DateTime, Boolean
+from sqlalchemy import Column, Integer, DateTime
 
 Base = declarative_base()
 
@@ -8,6 +8,6 @@ class BaseEntity(Base):
     
     __abstract__ = True
     
-    id = Column(BigInteger, primary_key=True, autoincrement=True, index=True)
+    id = Column(Integer, primary_key=True, autoincrement=True, index=True)
     created_at = Column(DateTime, default=datetime.now())
     updated_at = Column(DateTime)
